configured the project to work with typescript
