+++ src/lifters/MapSubjects.ts
@@ -71,7 +71,7 @@ export default class MapSubjects {
       const urlsElementsClassifier = {
         cssSelector: 'td[width=51].nombre > span.grupo',
         html: html,
-        filter: element => {
+        filter: (element: CheerioElement): string => {
             /*
           let generic_link = element.children[1].attribs.href;
             commented while the FCA website puts the links again
