+++ src/normalizers/Tagger.ts
@@ -1,7 +1,7 @@
 import Module from '../entities/Module';
 import Activity from '../entities/Activity';
 import Modules from '../entities/groupers/Modules';
-
+import {Block} from '../types';
 
 export default class Tagger {
 
@@ -44,7 +44,7 @@ export default class Tagger {
      * @param {String} moduleID
      * @param {[]} activities
      */
-    getSeparatedActivities(moduleID: string, activities: {text: string, actividad: string}[]): Activity[] {
+    getSeparatedActivities(moduleID: string, activities: Block[]): Activity[] {
         return activities.map( (act, actIndex: number) => {
             return new Activity(
                 moduleID,
